Reset softnav_painted_area_ when a new soft nav is about to happen
This is reset the softnav_painted_area_ when a new potential soft nav
is about to happen.

The image used in dropped-entries.tentative.html is changed for a
small one. That test adds the image 52 times to verify that entries
are dropped once the buffer is full. Small image is less like to
cause timeout which is seen in one dry run. The timeout is not related
to the change in this CL.

Change-Id: I6bff75e1b2e32d35fccc16b6b519eca66bf0ac59