[bfcache] Update NotRestoredReasons to match the spec draft
This CL updates NotRestoredReasons API's implementation to conform to the spec draft[1].
Namely, this CL
(1) removes "preventedBFCache" enum,
(2) adds "masked" to the outermost main frame when cross-origin iframes are blocking, and
(3) creates NotRestoredReasonDetails class on the blink side to make the reason into a struct like {"reason": "error-document"}.

As for (2), when `kAllowCrossOriginSubframes` flag is on, we add "masked" to the randomly selected cross-origin iframe.
Also with (2), `reasons` field is no longer specific to same-origin frames, so I'm moving the field out of `SameOriginDetails`.

[1]: https://github.com/whatwg/html/pull/9360/files

Bug: 1472637

Change-Id: I89702696a128eea31b2dd2af0ef8c9874e7213ca