Preserve opened layer content across canvas flushes
When the canvas has to be presented while layers are opened, we cannot
present the layer content, but we can present any painting that happened
before the top most layer was opened. To do so, the recording has to be
split into two parts: a main recording and a side recording. The side
recording holds the content of all nested layers in one single
recording. When the top most layer is closed, the content of the side
recording is written to the main recording.

Bug: 1484741
Change-Id: Id32412819c0395edd501243b2ccead30ebc0b674