Remove CSS Zoom Adjustment for GetBoundingClientRect
This change modifies the behavior of the call for elements with a CSS
property. Previously, the dimensions of the bounding box returned by
GetBoundingClientRect were adjusted to the effective zoom of an
element, which is the combination of the page zoom and the CSS zoom.
This meant that the dimensions were not affected by either of those
values.

This had the unintended consequence of causing two elements with the
same size before zoom, but with different zoom properties, to have the
same sized bounding box, even though they had different sizes on the
page. This also caused crbug.com/967097

The new behavior adjusts the bounding box to the page's zoom, but not to
the CSS zoom. This means that two elements with the same size before
zoom, but with different zoom properties, will now have different sized
bounding boxes.

Follow up patches will add the same logic to other calls in the file to
ensure consistency between the calls.

Bug: 967097, 1510969, 1511048

Change-Id: I66e8d4ca02f117c6133bb15cea29fc5c50c5d819